Fix SuperResolution implementation
Added an option to run SuperRes for all videos.
Decoupled SuperRes from VPScaling so users have more control over what shaders they can use, and not limiting video to D3D scaling when SuperRes is not enabled.
Implemented a better SuperRes check to remove more edge cases where the option is turned on, but SuperRes is not working.
SuperRes now plays nicer with HDR functionality.
User will now be better informed when SuperRes is enabled or not from the debug menu.

Might need testing for systems that are not Nvidia (i.e, Intel's SuperRes feature).